Личный кабинет владельца: CRUD питомцев

+++ src/main/java/sample/vet/Owner.java
@@ -1,13 +1,13 @@
 package sample.vet;
 
 public class Owner extends User{
-    private long owner_id;
+    private int owner_id;
     private String last_name;
     private String first_name;
     private String second_name;
     private String address;
     private String phone_number;
-    public Owner(long owner_id, String last_name, String first_name, String second_name, String address, String phone_number, String login, String password){
+    public Owner(int owner_id, String last_name, String first_name, String second_name, String address, String phone_number, String login, String password){
         super(login, password, 2);
         this.owner_id = owner_id;
         this.last_name = last_name;
@@ -26,7 +26,7 @@ public class Owner extends User{
     }
 
 
-    public long getOwner_id() {
+    public int getOwner_id() {
         return owner_id;
     }
 
@@ -50,7 +50,7 @@ public class Owner extends User{
         return phone_number;
     }
 
-    public void setOwner_id(long owner_id) {
+    public void setOwner_id(int owner_id) {
         this.owner_id = owner_id;
     }
